+++ Tests/SentryTests/Integrations/WatchdogTerminations/SentryWatchdogTerminationsIntegrationTests.swift
@@ -1,6 +1,6 @@
 import XCTest
 
-class SentryOutOfMemoryIntegrationTests: XCTestCase {
+class SentryWatchdogTerminationIntegrationTests: XCTestCase {
 
     private class Fixture {
         let options: Options
@@ -26,7 +26,7 @@ class SentryOutOfMemoryIntegrationTests: XCTestCase {
     }
     
     private var fixture: Fixture!
-    private var sut: SentryOutOfMemoryTrackingIntegration!
+    private var sut: SentryWatchdogTerminationTrackingIntegration!
     
     override func setUp() {
         super.setUp()
@@ -43,14 +43,14 @@ class SentryOutOfMemoryIntegrationTests: XCTestCase {
     }
     
     func testWhenUnitTests_TrackerNotInitialized() {
-        let sut = SentryOutOfMemoryTrackingIntegration()
+        let sut = SentryWatchdogTerminationTrackingIntegration()
         sut.install(with: Options())
         
         XCTAssertNil(Dynamic(sut).tracker.asAnyObject)
     }
     
     func testWhenNoUnitTests_TrackerInitialized() {
-        let sut = SentryOutOfMemoryTrackingIntegration()
+        let sut = SentryWatchdogTerminationTrackingIntegration()
         Dynamic(sut).setTestConfigurationFilePath(nil)
         sut.install(with: Options())
         
@@ -58,7 +58,7 @@ class SentryOutOfMemoryIntegrationTests: XCTestCase {
     }
     
     func testTestConfigurationFilePath() {
-        let sut = SentryOutOfMemoryTrackingIntegration()
+        let sut = SentryWatchdogTerminationTrackingIntegration()
         let path = Dynamic(sut).testConfigurationFilePath.asString
         XCTAssertEqual(path, ProcessInfo.processInfo.environment["XCTestConfigurationFilePath"])
     }
@@ -93,9 +93,9 @@ class SentryOutOfMemoryIntegrationTests: XCTestCase {
     func test_OOMDisabled_RemovesEnabledIntegration() {
         givenInitializedTracker()
         let options = Options()
-        options.enableOutOfMemoryTracking = false
+        options.enableWatchdogTerminationTracking = false
         
-        let sut = SentryOutOfMemoryTrackingIntegration()
+        let sut = SentryWatchdogTerminationTrackingIntegration()
         let result = sut.install(with: options)
         
         XCTAssertFalse(result)
@@ -103,7 +103,7 @@ class SentryOutOfMemoryIntegrationTests: XCTestCase {
     
     private func givenInitializedTracker(isBeingTraced: Bool = false) {
         fixture.crashWrapper.internalIsBeingTraced = isBeingTraced
-        sut = SentryOutOfMemoryTrackingIntegration()
+        sut = SentryWatchdogTerminationTrackingIntegration()
         let options = Options()
         Dynamic(sut).setTestConfigurationFilePath(nil)
         sut.install(with: options)